requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

googletrans
keyboard
prompt-toolkit
rich
spotipy

questionary
ytmusicapi
PyYAML
sqlitedict
textual
pillow